refactor(catalog): extract apis extensions file
